Remove inactive committers
The TSC decided to remove the committers that have been inactive in gerit
for a year or more. If any of the people removed by this patch are contributing
to the project in other ways, and the PTL would like to keep the gerrit rights
for them, please ammend this patch.

For other updates on the committer list, please follow the procedure descrbed in
https://wiki.opnfv.org/display/DEV/Developer+governance+and+reports

Updated the PTL per the new PTL's request

JIRA: TSC-14

Change-Id: I28502ba4ade4d5e2863366f358b9f1700235785d
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>